+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-type-select/src/main/resources/META-INF/resources/select_field.js
@@ -125,6 +125,16 @@ AUI.add(
 							value = inputNode.val();
 						}
 
+						if (!value) {
+							var contextValue = instance._getContextValue();
+
+							var hasOption = instance._hasOption(contextValue);
+
+							if (contextValue && !hasOption) {
+								value = contextValue;
+							}
+						}
+
 						return value;
 					},
 
@@ -195,6 +205,10 @@ AUI.add(
 								instance._setSelectNodeOptions(optionNode, value);
 							}
 						);
+
+						instance.set('value', value);
+
+						instance.render();
 					},
 
 					showErrorMessage: function() {
@@ -233,6 +247,18 @@ AUI.add(
 						}
 					},
 
+					_getContextValue: function() {
+						var instance = this;
+
+						var contextValue = instance.get('value');
+
+						if (Lang.isArray(contextValue)) {
+							contextValue = contextValue[0];
+						}
+
+						return contextValue;
+					},
+
 					_getDataSourceType: function(value) {
 						if (Lang.isString(value)) {
 							try {
@@ -265,7 +291,7 @@ AUI.add(
 								function(value, index) {
 									options.forEach(
 										function(option, index) {
-											if (option.value.indexOf(value) > -1) {
+											if (value && option.value === value) {
 												optionsSelected.push(option);
 											}
 										}
@@ -283,6 +309,24 @@ AUI.add(
 						return instance.get('container').one('.' + CSS_SELECT_TRIGGER_ACTION);
 					},
 
+					_hasOption: function(value) {
+						var instance = this;
+
+						var hasOption = false;
+
+						var inputNode = instance.getInputNode();
+
+						inputNode.all('option').each(
+							function(optionNode) {
+								if (optionNode.val() === value) {
+									hasOption = true;
+								}
+							}
+						);
+
+						return hasOption;
+					},
+
 					_isClickingOutSide: function(event) {
 						var instance = this;
 
@@ -296,6 +340,10 @@ AUI.add(
 
 						var container = instance.get('container');
 
+						if (!container.one('.drop-chosen')) {
+							return false;
+						}
+
 						var openList = container.one('.drop-chosen').hasClass('hide');
 
 						return !openList;
@@ -304,15 +352,9 @@ AUI.add(
 					_onClickItem: function(event) {
 						var instance = this;
 
-						var options = instance.get('options');
-
 						var value = event.target.getAttribute('data-option-value');
 
 						instance.setValue(value);
-
-						instance.set('value', [value]);
-
-						instance.render();
 					},
 
 					_selectDOMOption: function(optionNode, value) {