+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/Dom4JUtil.java
@@ -71,25 +71,36 @@ public class Dom4JUtil {
 
 		Writer writer = new CharArrayWriter();
 
-		XMLWriter xmlWriter = pretty ? new XMLWriter(
-			writer, OutputFormat.createPrettyPrint()) : new XMLWriter(writer);
+		OutputFormat outputFormat = OutputFormat.createPrettyPrint();
+
+		outputFormat.setTrimText(false);
+
+		XMLWriter xmlWriter = null;
+
+		if (pretty) {
+			xmlWriter = new XMLWriter(writer, outputFormat);
+		}
+		else {
+			xmlWriter = new XMLWriter(writer);
+		}
 
 		xmlWriter.write(element);
 
-		return writer.toString();
+		String formattedContent = writer.toString();
+
+		return formattedContent.replace("<code> ", "<code>\n ");
 	}
 
 	public static Element getNewAnchorElement(
 		String href, Element parentElement, Object... items) {
 
+		if ((items == null) || (items.length == 0)) {
+			return null;
+		}
+
 		Element anchorElement = null;
 
-		if (parentElement == null) {
-			anchorElement = new DefaultElement("a");
-		}
-		else {
-			anchorElement = getNewElement("a", parentElement);
-		}
+		anchorElement = getNewElement("a", parentElement);
 
 		anchorElement.addAttribute("href", href);
 
@@ -102,12 +113,18 @@ public class Dom4JUtil {
 		return getNewAnchorElement(href, null, items);
 	}
 
+	public static Element getNewElement(String childElementTag) {
+		return getNewElement(childElementTag, null);
+	}
+
 	public static Element getNewElement(
 		String childElementTag, Element parentElement, Object... items) {
 
 		Element childElement = new DefaultElement(childElementTag);
 
-		parentElement.add(childElement);
+		if (parentElement != null) {
+			parentElement.add(childElement);
+		}
 
 		if ((items != null) && (items.length > 0)) {
 			addToElement(childElement, items);
@@ -164,30 +181,10 @@ public class Dom4JUtil {
 	}
 
 	public static Element toCodeSnippetElement(String content) {
-		return wrapWithNewElement(
-			wrapWithNewElement(
-				JenkinsResultsParserUtil.redact(content), "code"),
-			"pre");
-	}
-
-	public static Element wrapWithNewElement(
-		Element element, String wrapperTag) {
-
-		Element wrapperElement = new DefaultElement(wrapperTag);
-
-		wrapperElement.add(element);
-
-		return wrapperElement;
-	}
-
-	public static Element wrapWithNewElement(
-		String content, String wrapperTag) {
-
-		Element wrapperElement = new DefaultElement(wrapperTag);
-
-		wrapperElement.addText(content);
-
-		return wrapperElement;
+		return getNewElement(
+			"pre", null,
+			getNewElement(
+				"code", null, JenkinsResultsParserUtil.redact(content)));
 	}
 
 }