+++ modules/util/source-formatter/src/main/java/com/liferay/source/formatter/BNDSourceProcessor.java
@@ -250,6 +250,13 @@ public class BNDSourceProcessor extends BaseSourceProcessor {
 
 		content = formatBundleClassPath(content);
 
+		int pos = fileName.lastIndexOf(StringPool.SLASH);
+
+		String shortFileName = fileName.substring(pos + 1);
+
+		content = formatLineBreaks(shortFileName, content);
+		content = formatWhitespace(shortFileName, content);
+
 		if ((portalSource || subrepository) && isModulesFile(absolutePath) &&
 			!fileName.endsWith("test-bnd.bnd")) {
 
@@ -381,6 +388,73 @@ public class BNDSourceProcessor extends BaseSourceProcessor {
 
 		return sortDefinitionProperties(
 			content, includeResources, new IncludeResourceComparator());
+	}
+
+	protected String formatLineBreaks(
+		String content, Map<String, String> definitionsKeysMap) {
+
+		if (definitionsKeysMap == null) {
+			return content;
+		}
+
+		for (Map.Entry<String, String> entry : definitionsKeysMap.entrySet()) {
+			String definitionKey = entry.getValue();
+
+			Pattern pattern = Pattern.compile(" " + definitionKey + ":");
+
+			Matcher matcher = pattern.matcher(content);
+
+			if (matcher.find()) {
+				return StringUtil.replaceFirst(
+					content, " ", "\n", matcher.start());
+			}
+		}
+
+		return content;
+	}
+
+	protected String formatLineBreaks(String shortFileName, String content) {
+		content = formatLineBreaks(content, getDefinitionKeysMap());
+
+		Map<String, Map<String, String>> fileSpecificDefinitionKeysMap =
+			getFileSpecificDefinitionKeysMap();
+
+		return formatLineBreaks(
+			content, fileSpecificDefinitionKeysMap.get(shortFileName));
+	}
+
+	protected String formatWhitespace(
+		String content, Map<String, String> definitionsKeysMap) {
+
+		if (definitionsKeysMap == null) {
+			return content;
+		}
+
+		for (Map.Entry<String, String> entry : definitionsKeysMap.entrySet()) {
+			String definitionKey = entry.getValue();
+
+			Pattern pattern = Pattern.compile(
+				"(\\A|\n)" + definitionKey + ":[^ \\\\\n]");
+
+			Matcher matcher = pattern.matcher(content);
+
+			if (matcher.find()) {
+				return StringUtil.insert(
+					content, StringPool.SPACE, matcher.end() - 1);
+			}
+		}
+
+		return content;
+	}
+
+	protected String formatWhitespace(String shortFileName, String content) {
+		content = formatWhitespace(content, getDefinitionKeysMap());
+
+		Map<String, Map<String, String>> fileSpecificDefinitionKeysMap =
+			getFileSpecificDefinitionKeysMap();
+
+		return formatWhitespace(
+			content, fileSpecificDefinitionKeysMap.get(shortFileName));
 	}
 
 	protected Map<String, String> getDefinitionKeysMap() {